Migrate ramps to ensure validity
We added validation checks to make sure that ramp stops are sorted by
zoom level and do not have duplicate stops by zoom. This broke some
styles that have ramps with out of order and/or duplicate stops.

We should migrate these styles ramp to ensure compliance with the new
validations. The migration does the following:

- sorts ramp stops by zoom
- removes duplicate stops
- marginally increments the zoom of a colliding stop with a unique value
  - rerun the migration in case changing the zoom causes a new
    collision/ordering issue